refactor(ui): promote MentionDropdown from RichTextEditor internal to shared ui component
